Move position types to position module.
